Refactor sync - renamed files from ee to cloud
