+++ agenta-backend/agenta_backend/__init__.py
@@ -1,4 +1,4 @@
 import os
 
 if os.environ["FEATURE_FLAG"] in ["cloud", "ee", "demo"]:
-    import agenta_backend.ee.__init__
+    import agenta_backend.cloud.__init__
