+++ agenta-backend/agenta_backend/routers/app_router.py
@@ -28,7 +28,7 @@ from agenta_backend.models.api.api_models import (
 from agenta_backend.models import converters
 
 if os.environ["FEATURE_FLAG"] in ["cloud", "ee", "demo"]:
-    from agenta_backend.ee.services.selectors import (
+    from agenta_backend.cloud.services.selectors import (
         get_user_and_org_id,
     )  # noqa pylint: disable-all
 else:
