+++ agenta-backend/agenta_backend/routers/configs_router.py
@@ -13,7 +13,7 @@ from agenta_backend.services import (
 )
 
 if os.environ["FEATURE_FLAG"] in ["cloud", "ee", "demo"]:
-    from agenta_backend.ee.services.selectors import (
+    from agenta_backend.cloud.services.selectors import (
         get_user_and_org_id,
     )  # noqa pylint: disable-all
 else:
