+++ agenta-backend/agenta_backend/routers/organization_router.py
@@ -13,7 +13,7 @@ from agenta_backend.models.api.organization_models import (
 from agenta_backend.services import db_manager
 
 if os.environ["FEATURE_FLAG"] in ["cloud", "ee", "demo"]:
-    from agenta_backend.ee.services.selectors import (
+    from agenta_backend.cloud.services.selectors import (
         get_user_and_org_id,
     )  # noqa pylint: disable-all
 else:
