+++ agenta-backend/agenta_backend/services/app_manager.py
@@ -17,14 +17,14 @@ from agenta_backend.models.db_models import (
 from agenta_backend.services import db_manager
 
 if os.environ["FEATURE_FLAG"] in ["cloud"]:
-    from agenta_backend.ee.services import (
+    from agenta_backend.cloud.services import (
         lambda_deployment_manager as deployment_manager,
     )  # noqa pylint: disable-all
 else:
     from agenta_backend.services import deployment_manager
 
 if os.environ["FEATURE_FLAG"] in ["cloud", "ee", "demo"]:
-    from agenta_backend.ee.services import (
+    from agenta_backend.cloud.services import (
         api_key_service,
     )  # noqa pylint: disable-all
